+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-03-10T17:52:40.977Z",
+    "2026-03-11T09:45:14.163Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {